pyproject.toml
[project]
name = "pystudy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dashscope>=1.25.4",
    "deepagents>=0.3.0",
    "gradio>=6.1.0",
    "langchain>=1.1.3",
    "langchain-community>=0.4.1",
    "langchain-deepseek>=1.0.1",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.1",
    "langchain-postgres>=0.0.16",
    "python-dotenv>=1.2.1",
]
